services/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
pymilvus>=2.3.4
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
Pillow==10.0.1
structlog==23.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
email-validator==2.1.0
prometheus-client==0.19.0